summary cleanup
