+++ plugins/summary/src/fromTemplate.js
